+++ pendulum/tz/__init__.py
@@ -21,6 +21,24 @@ def local_timezone():
     """
     Loads the local timezone.
 
-    :rtype: Timezone
+    :rtype: LocalTimezone
     """
     return LocalTimezone.get()
+
+
+def test_local_timezone(mock):
+    """
+    Context manager to temporarily set the local_timezone value.
+
+    :type mock: pendulum.tz.timezone.Timezone or str
+    """
+    return LocalTimezone.test(mock)
+
+
+def set_local_timezone(timezone=None):
+    """
+    Set the the local timezone to a given one.
+
+    :type mock: pendulum.tz.timezone.Timezone or str or None
+    """
+    LocalTimezone.set(timezone)